Fixed MonitorTrigger class redefinition
